Extract Cocktail component
